Rename lookupTables to crcLookupTables.
