+++ server/udp/crcLookupTables.js
